+++ packages/widgetbook/test/generator/custom_args/custom_args.stories.g.dart
@@ -23,6 +23,7 @@ typedef LabelBadgeDefaults = Defaults<LabelBadge, NumericBadgeInputArgs>;
 class LabelBadgeStory extends Story<LabelBadge, NumericBadgeInputArgs> {
   LabelBadgeStory({
     super.name,
+    super.designLink,
     super.setup,
     super.modes,
     NumericBadgeInputArgs? args,