+++ packages/widgetbook/test/generator/doc_comment_code/doc_comment_code.stories.g.dart
@@ -32,6 +32,7 @@ class DocCommentCodeWidgetStory
     extends Story<DocCommentCodeWidget, DocCommentCodeWidgetArgs> {
   DocCommentCodeWidgetStory({
     super.name,
+    super.designLink,
     super.setup,
     super.modes,
     DocCommentCodeWidgetArgs? args,