+++ packages/widgetbook/test/generator/enums/enums.stories.g.dart
@@ -23,6 +23,7 @@ typedef EnumWidgetDefaults = Defaults<EnumWidget, EnumWidgetArgs>;
 class EnumWidgetStory extends Story<EnumWidget, EnumWidgetArgs> {
   EnumWidgetStory({
     super.name,
+    super.designLink,
     super.setup,
     super.modes,
     EnumWidgetArgs? args,