+++ packages/widgetbook/test/generator/generic/generic.stories.g.dart
@@ -26,6 +26,7 @@ class GenericWidgetStory<T extends num>
     extends Story<GenericWidget<T>, GenericWidgetArgs<T>> {
   GenericWidgetStory({
     super.name,
+    super.designLink,
     super.setup,
     super.modes,
     required super.args,